+++ srcs/sam/rays.c
@@ -6,7 +6,7 @@
 /*   By: saperrie <saperrie@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/15 19:55:31 by saperrie          #+#    #+#             */
-/*   Updated: 2024/10/15 22:48:19 by saperrie         ###   ########.fr       */
+/*   Updated: 2024/10/21 14:22:27 by saperrie         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 int	raycast(char **map, t_player *player, t_pplane *pplane)
 {
-	float	player_position;
-	float	ray;
+	double	player_position;
+	double	ray;
 	int		columns;
